src/train.py
"""
Training script for object detection models
"""

import argparse
import json
import logging
from pathlib import Path

import torch
import yaml
from torch.utils.data import DataLoader

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
)
logger = logging.getLogger(__name__)


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    with open(config_path, "r") as f:
        config = yaml.safe_load(f)
    return config


def setup_device(device_type: str = "cuda") -> torch.device:
    """Setup training device."""
    if device_type == "cuda" and torch.cuda.is_available():
        device = torch.device("cuda")
        logger.info(f"Using CUDA: {torch.cuda.get_device_name(0)}")
    else:
        device = torch.device("cpu")
        logger.info("Using CPU")
    return device


def main(args):
    """Main training function."""
    # Load configuration
    config = load_config(args.config)
    logger.info(f"Loaded configuration from {args.config}")
    logger.info(f"Configuration: {json.dumps(config, indent=2)}")

    # Setup device
    device = setup_device(config["device"]["type"])

    # Create output directories
    log_dir = Path(config["logging"]["log_dir"])
    checkpoint_dir = Path(config["logging"]["checkpoint_dir"])
    log_dir.mkdir(parents=True, exist_ok=True)
    checkpoint_dir.mkdir(parents=True, exist_ok=True)

    logger.info("Training configuration:")
    logger.info(f"  - Batch size: {config['training']['batch_size']}")
    logger.info(f"  - Learning rate: {config['training']['learning_rate']}")
    logger.info(f"  - Epochs: {config['training']['num_epochs']}")
    logger.info(f"  - Model: {config['model']['type']}")

    # TODO: Implement data loading
    # TODO: Implement model initialization
    # TODO: Implement training loop
    # TODO: Implement validation loop
    # TODO: Implement checkpointing
    # TODO: Implement logging to TensorBoard

    logger.info("Training setup complete. Awaiting implementation of training loop.")


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Train object detection model")
    parser.add_argument(
        "--config",
        type=str,
        default="config/config.yaml",
        help="Path to configuration file",
    )
    parser.add_argument(
        "--resume",
        type=str,
        default=None,
        help="Path to checkpoint to resume training",
    )
    parser.add_argument(
        "--output-dir", type=str, default="outputs", help="Output directory"
    )

    args = parser.parse_args()
    main(args)
