requirements.txt
# Core ML/DL frameworks
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Data processing and numerical
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Computer vision (note: opencv-python-headless for container)
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
tensorboard==2.15.1

# Jupyter and interactive
jupyter==1.0.0
jupyterlab==4.0.9
ipython==8.18.1
ipywidgets==8.1.1

# Development tools
pytest==7.4.3
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2

# Data loading and preprocessing
scikit-image==0.22.0
scikit-learn==1.3.2
albumentations==1.3.1

# Additional utilities
tqdm==4.66.1
pyyaml==6.0.1
dotenv==0.21.0
python-dotenv==1.0.0

# Optional: for model deployment and serving
onnx==1.15.0
onnxruntime==1.17.0

# Optional: for experiment tracking
wandb==0.16.1
